small updates and some documentation

+++ parser/fetch_packages.py
@@ -20,6 +20,7 @@ pypi_session = requests.sessions.Session()
 
 
 def get_best_package(package_name, specifier=''):
+    print(f'get_best_package: {package_name}')
     allow_yanked = True
     ignore_requires_python = True
     target_python = TargetPython(py_version_info=(3, 8, 3))
@@ -40,22 +41,27 @@ def get_best_package(package_name, specifier=''):
 
 
 def download_best_package(package_name, specifier='', path=PACKAGES_DIR):
+    print(f'download_best_package: {package_name}')
     name, version, url = get_best_package(package_name, specifier)
     url_split = url.split('#sha256')[0]
     file_split = url_split.split('/')[-1]
+    dest = f'{path}/{name}/{file_split}'
 
     if not os.path.exists(f'{path}/{name}'):
-       os.makedirs(f'{path}/{name}')
+        os.makedirs(f'{path}/{name}')
 
-    dest = f'{path}/{name}/{file_split}'
-    r = requests.get(url_split, allow_redirects=True)
-    open(dest, 'wb').write(r.content)
-    shutil.unpack_archive(dest, f'{path}/{package_name}')
-    extracted = dest.rstrip('.tar.gz')
-    os.rename(extracted, extracted.lower())
+    # Only download in not done previously
+    if not os.path.isfile(dest):
+        print(f'Downloading: {url_split}')
+        r = requests.get(url_split, allow_redirects=True)
+        open(dest, 'wb').write(r.content)
+        shutil.unpack_archive(dest, f'{path}/{package_name}')
+        extracted = dest.rstrip('.tar.gz')
+        os.rename(extracted, extracted.lower())
 
 
 def get_package_info(package):
+    print(f'get_package_info: {package}')
     url = 'https://pypi.python.org/pypi/' + package + '/json'
     response = pypi_session.get(url, allow_redirects=True)
     data = response.json()
@@ -64,6 +70,7 @@ def get_package_info(package):
 
 # fetch requirements.txt
 def fetch_all_from_source():
+    print("### In fetch_all_from_source ###")
     resp = requests.get(SOURCE_URL)
     pkgs = resp.text
 
@@ -79,6 +86,7 @@ def fetch_all_from_source():
 
 
 def fetch_pkg_dependencies(pkg_name):
+    print("### In fetch_pkg_dependencies ###")
     fetched = get_package_info(pkg_name)  # fetch from pypi
     dependencies = fetched['info']['requires_dist']
     to_return = []
@@ -100,6 +108,7 @@ def fetch_pkg_dependencies(pkg_name):
 
 
 def fetch_all_inc_deps(all_packages):
+    print("### In fetch_all_inc_deps ###")
     work_queue = deque()
     known_packages = set()  # Packages we all-ready know
     for pkg in all_packages:
@@ -109,8 +118,8 @@ def fetch_all_inc_deps(all_packages):
     fetched_list = []
     while work_queue:
         pkg = work_queue.popleft()
-        deps = fetch_pkg_dependencies(pkg['name'])
         print(f"Fetching {pkg['name']}")
+        deps = fetch_pkg_dependencies(pkg['name'])
         fetched_list.append({
             **pkg,
             'dependencies': deps
@@ -125,6 +134,7 @@ def fetch_all_inc_deps(all_packages):
 
 
 def _condense_dependencies(fetched_list, facit):
+    print("### In _condense_dependencies ###")
     packages = {}
     for pkg in facit:  # add known packages with definite versions to dict
         packages[pkg['name']] = {'name': pkg['name'], 'definite_version': pkg['version'], 'required_by': {}}
@@ -185,17 +195,18 @@ def download_all_packages(condensed=None):
 
 
 if __name__ == '__main__':
+    # download_all_packages('out.json')
+    res = fetch_all_requirements_reqs()
+
+    with open('fetched_all.json', 'w') as fp:
+        json.dump(res, fp, indent=4)
+
+    with open('fetched_all.json', 'r') as fp:
+        loaded = fp.read()
+        fetched_list = json.loads(loaded.lower())
+        _set_definite_versions(fetched_list)
+        with open('out.json', 'w') as out:
+            json.dump(fetched_list, out, indent=4)
+
     download_all_packages('out.json')
-    # res = fetch_all_requirements_reqs()
 
-    # with open('fetched_all.json', 'w') as fp:
-    #     json.dump(res, fp, indent=4)
-
-    # facit = fetch_all_from_source()
-    # with open('fetched_all.json', 'r') as fp:
-    #     loaded = fp.read()
-    #     fetched_list = json.loads(loaded.lower())
-    #     condensed = _condense_dependencies(fetched_list, facit)
-    #     set_definite_versions(condensed)
-    #     with open('out.json', 'w') as out:
-    #         json.dump(condensed, out, indent=4)
